Remove cache for /glist
We cannot guarantee that a cached display is still reflecting a real view of players in the proxy, including vanished players.

In such case, a cache integrity check should be implemented to verify if cache is still valid or not, but to do that, we need to iterate all players connected to the proxy and re-calculate every server, so in terms of performance, implementing a cache integrity check with have the same cost as if /glist is calculated on every execution without a cache.